add weather component onto static-page
