+++ my-app/src/App.js
@@ -21,7 +21,6 @@ import Allplans from './pages/AllPlans';
 import firebaseDB from './utils/firebaseConfig';
 import { getDocs, getDoc, collection, doc } from 'firebase/firestore';
 import Weather from './components/weather/Weather';
-import WeatherCom from './components/weather/WeatherCom';
 
 const db = firebaseDB();
 
@@ -72,7 +71,7 @@ function App() {
       <Header />
       <BodyWrapper>
         <Routes>
-          <Route path="/" element={<WeatherCom />} />
+          <Route path="/" element={<Weather />} />
           <Route
             path="/edit-plan-detail"
             element={
